+++ django/applications/catmaid/static/js/tools/edit-tool.js
@@ -139,18 +139,6 @@
         }));
     }
 
-    editToolActions.push(new CATMAID.Action({
-      helpText: "Open widget",
-      keyShortcuts: {
-        'F6': [ 117 ]
-      },
-      run: function (e) {
-        var dialog = new CATMAID.OpenWidgetDialog();
-        dialog.show();
-        return true;
-      }
-    }));
-
     return editToolActions;
   };